% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{addDataCols}
\alias{addDataCols}
\title{addDataCols}
\usage{
addDataCols(data = NULL, de_res = NULL, cols = NULL, conds = NULL)
}
\arguments{
\item{data}{data}

\item{de_res}{DE results}

\item{cols}{columns}

\item{conds}{conditions}
}
\description{
add data columns to de results. Adapted from debrowser::addDataCols().
}
\examples{
     x <- addDataCols()
     
}
