% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{applyFiltersIter}
\alias{applyFiltersIter}
\title{applyFiltersIter}
\usage{
applyFiltersIter(data = NULL, input = NULL)
}
\arguments{
\item{data}{loaded dataset}

\item{input}{input parameters}
}
\description{
Apply filter for Initial and Final DEgenes
}
\examples{
     x <- applyFiltersIter()

}
