% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{connectMongo}
\alias{connectMongo}
\title{connectMongo}
\usage{
connectMongo(input = NULL, database = NULL, host = NULL)
}
\arguments{
\item{input}{input}

\item{database}{database}

\item{host}{host}
}
\description{
connect to a Mongo Server
}
\examples{
      mongo_connection <- connectMongo(input)
    
}
