% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{cutOffSelectionUI}
\alias{cutOffSelectionUI}
\title{cutOffSelectionUI}
\usage{
cutOffSelectionUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\description{
Gathers the cut off selection for DE analysis
}
\examples{
    x <- cutOffSelectionUI("cutoff")
    
}
