% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataLoadServer}
\alias{dataLoadServer}
\title{dataLoadServer}
\usage{
dataLoadServer(
  input = NULL,
  output = NULL,
  session = NULL,
  nextpagebutton = NULL,
  parent_session = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output objects}

\item{session}{session}

\item{nextpagebutton}{the name of the next page button after loading the data}

\item{parent_session}{the parent session}
}
\value{
panel
}
\description{
Data loading module for bulk and single cell data sets.
}
