% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataProfileSummaryUI}
\alias{dataProfileSummaryUI}
\title{dataProfileSummaryUI}
\usage{
dataProfileSummaryUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\description{
Creates data table and figures to summarize uploaded data
}
\examples{
    x <- dataProfileSummaryUI("load")

}
