% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataSummaryUI}
\alias{dataSummaryUI}
\title{dataSummaryUI}
\usage{
dataSummaryUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
Creates a panel to view the summary of uploaded data
}
\examples{
    x <- dataLoadUI("load")

}
