% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataUI}
\alias{dataUI}
\title{dataLoadUI}
\usage{
dataUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Creates panels to upload the data. Adapted from debrowser::dataLoadUI().
}
\examples{
    x <- dataLoadUI("load")
    
}
