% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{display_venn}
\alias{display_venn}
\title{display_venn}
\usage{
display_venn(x, category.names, ...)
}
\arguments{
\item{x}{a list of elements for each groups}

\item{...}{}
}
\description{
display venn diagram
}
