% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalUI.R
\name{dprofilerbarmainplot}
\alias{dprofilerbarmainplot}
\title{dprofilerbarmainplot}
\usage{
dprofilerbarmainplot(
  input = NULL,
  output = NULL,
  session = NULL,
  data = NULL,
  cols = NULL,
  conds = NULL,
  key = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output variables}

\item{session}{session}

\item{data}{a matrix with expression values}

\item{cols}{columns}

\item{conds}{conditions}

\item{key}{the gene or region name}
}
\description{
Module for a bar plot that can be used in data prep, main plots low count removal modules or any desired module. 
Adapted from dprofilerbarmainplot().
}
\examples{
    x <- dprofilerbarmainplot()
    
}
