% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalServers.R
\name{dprofilerBatch}
\alias{dprofilerBatch}
\title{dprofilerBatch}
\usage{
dprofilerBatch(input, output, session, parent_session = NULL, filtd = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{parent_session, }{parent session}

\item{filtd, }{filtered data}
}
\value{
panel
}
\description{
Module to correct batch effect, adapted from debrowser::debrowserbatcheffect().
}
\examples{
    x <- dprofilerBatch()

}
