% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalUI.R
\name{dprofilerboxmainplot}
\alias{dprofilerboxmainplot}
\title{dprofilerboxmainplot}
\usage{
dprofilerboxmainplot(
  input = NULL,
  output = NULL,
  session = NULL,
  data = NULL,
  cols = NULL,
  conds = NULL,
  key = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output variables}

\item{session}{session}

\item{data}{a matrix with expression values}

\item{cols}{columns}

\item{conds}{conditions}

\item{key}{the gene or region name}
}
\description{
Module for a box plot that can be used alongside DEanalysis and heatmaps. Adapted from debrowserboxmainplot().
}
\examples{
    x <- dprofilerboxmainplot()
    
}
