% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{dprofilercondselect}
\alias{dprofilercondselect}
\title{dprofilercondselect}
\usage{
dprofilercondselect(
  input = NULL,
  output = NULL,
  session = NULL,
  parent_session = NULL,
  data = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output objects}

\item{session}{session}

\item{datax}{bulk or single cell data}

\item{marker_table}{marker table of either reference scRNA or bulk RNA data}

\item{profiling}{if TRUE, condition selection is conducted for external data sets (default: FALSE)}
}
\description{
Condition selection for DE analysis and reference single cell data. Adapted from debrowsercondselect().
}
\examples{
    x <- dprofilercondselect()
    
}
