% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{dprofilerdeanalysis}
\alias{dprofilerdeanalysis}
\title{dprofilerdeanalysis}
\usage{
dprofilerdeanalysis(
  input = NULL,
  output = NULL,
  session = NULL,
  parent_session = NULL,
  cond_dc = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{parent_session}{parent session}

\item{cond_dc}{data and DE parameters}
}
\description{
Module to perform and visualize Iterative DE results.
}
\examples{
    x <- dprofilerdeanalysis()
    
}
