% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{dprofilerdeconvolute}
\alias{dprofilerdeconvolute}
\title{dprofilerdeconvolute}
\usage{
dprofilerdeconvolute(
  input = NULL,
  output = NULL,
  session = NULL,
  dc = NULL,
  scdata = NULL,
  parent_session = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{dc, }{de results}

\item{scdata, }{single cell data}

\item{parent_session}{main session}
}
\value{
DE panel
}
\description{
Module to perform and visualize deconvolution results.
}
\examples{
    x <- dprofilerdeconvolute()

}
