% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{dprofilerdeconvolute_old}
\alias{dprofilerdeconvolute_old}
\title{dprofilerdeconvolute}
\usage{
dprofilerdeconvolute_old(
  input = NULL,
  output = NULL,
  session = NULL,
  dc = NULL,
  scdata = NULL,
  sc_marker_table = NULL,
  parent_session = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output objects}

\item{session}{session}

\item{dc}{de results}

\item{scdata}{single cell data}

\item{sc_marker_table}{single cell marker table}

\item{parent_session}{parent session}
}
\value{
DE panel
}
\description{
Module to perform and visualize deconvolution results.
}
\examples{
    x <- dprofilerdeconvolute() 
    
}
