% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{dprofilerheatmap}
\alias{dprofilerheatmap}
\title{dprofilerheatmap}
\usage{
dprofilerheatmap(input, output, session, expdata = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{dc}{differential analysis data}

\item{selected_genes}{selected genes in the main plot}
}
\description{
Heatmap module to create interactive heatmaps and get selected list from
a heatmap
}
\examples{
    x <- dprofilerheatmap()
    
}
