% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{dprofilermainplot}
\alias{dprofilermainplot}
\title{dprofilermainplot}
\usage{
dprofilermainplot(input = NULL, output = NULL, session = NULL, data = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}
}
\description{
Module for a scatter, volcano and ma plots that are going to be used 
as a mainplot in dprofiler. Adapted from debrowser::debrowsermainplot()
}
\examples{
    x <- dprofilermainplot()
    
}
