% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{dprofilersummariser}
\alias{dprofilersummariser}
\title{summaryServer}
\usage{
dprofilersummariser(
  input = NULL,
  output = NULL,
  session = NULL,
  parent_session = NULL,
  comppheno = NULL,
  composprof = NULL,
  comparprof = NULL
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{comppheno}{results from computational phenotypic profiling}

\item{composprof}{results from cmopositional profiling}

\item{comparprof}{results from comparative profiling}
}
\description{
a module server to summarise all membership scores
}
\examples{
    x <- summaryServer()
    
}
