% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{dprofilerUI}
\alias{dprofilerUI}
\title{dprofilerUI}
\usage{
dprofilerUI()
}
\value{
the panel for main plots;
}
\description{
Creates a shinyUI to be able to run DEBrowser interactively.
}
\note{
\code{dprofilerUI}
}
\examples{
    x<-dprofilerUI()
    
}
