% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{external_silhouette}
\alias{external_silhouette}
\title{external_silhouette}
\usage{
external_silhouette(cluster = NULL, dist2 = NULL)
}
\arguments{
\item{cluster}{cluster labels of profile data}

\item{dist2}{non-symmetric similarity matrix}
}
\description{
Calculate silhouette measure for non-symmetric distance matrices. 
Suitable for comparison of expression profiles across different datasets and experiments
}
\examples{
     x <- external_silhouette()
    
}
