% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{fileSummaryBox}
\alias{fileSummaryBox}
\title{fileSummaryBox}
\usage{
fileSummaryBox(
  id = NULL,
  upload = NULL,
  sample = NULL,
  label = NULL,
  idbox = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{upload}{upload id}

\item{sample}{sample id}

\item{label}{label}

\item{idbox}{id of the shiny box}
}
\description{
file summary box for detailing the gene expression data and metadata
}
\examples{
      x <- fileSummaryBox("meta", "count", "metadata", "Metadata")
    
}
