% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{fileUploadBox}
\alias{fileUploadBox}
\title{fileUploadBox}
\usage{
fileUploadBox(
  id = NULL,
  inputId_count = NULL,
  inputId_meta = NULL,
  label = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{inputId_count}{input data file ID}

\item{inputId_meta}{input metadata file ID}

\item{label}{label}
}
\description{
file upload box for bulk data and metadata. Adapted from debrowser::fileUploadBox()
}
\examples{
      x <- fileUploadBox("meta", "count", "metadata", "Metadata")

}
