% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getAllMarkerGenes}
\alias{getAllMarkerGenes}
\title{getAllMarkerGenes}
\usage{
getAllMarkerGenes(scdata = NULL, data = NULL, input = NULL)
}
\arguments{
\item{scdata}{single cell data}

\item{data}{reference bulk data}

\item{input}{input}
}
\description{
getAllMarkerGenes
}
\examples{
    x <- getAllMarkerGenes()
    
}
