% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getAllProfGenes}
\alias{getAllProfGenes}
\title{getAllProfGenes}
\usage{
getAllProfGenes(prof_marker_table = NULL, data = NULL, input = NULL)
}
\arguments{
\item{prof_marker_table}{comparative profiling marker table}

\item{data}{submitted bulk data}

\item{input}{input}
}
\description{
getAllProfGenes
}
\examples{
    x <- getAllProfGenes()
    
}
