% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{getAllScoreDetails}
\alias{getAllScoreDetails}
\title{getAllScoreDetails}
\usage{
getAllScoreDetails(
  output = NULL,
  session = NULL,
  tablename = NULL,
  data = NULL,
  modal = NULL,
  highlight = FALSE
)
}
\arguments{
\item{output, }{output}

\item{session, }{session}

\item{tablename, }{table name}

\item{data, }{matrix data}

\item{modal, }{if it is true, the matrix is going to be in a modal}

\item{highlight}{if it is true, numerical columns are highlighted}
}
\description{
Details and scores of all profiling methods at the final Summary Pane
}
\examples{
    x <- getAllScoreDetails()
    
}
