% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getCCAEmbedding}
\alias{getCCAEmbedding}
\title{getCCAEmbedding}
\usage{
getCCAEmbedding(output = NULL, summary = NULL, data = NULL, input = NULL)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{data}{single cell ExpressionSet Object}

\item{input}{input}
}
\description{
get single cell RNA Embedding
}
\examples{
      x <- getCCAEmbedding()

}
