% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getDataAssesmentText}
\alias{getDataAssesmentText}
\title{getDataAssesmentText}
\usage{
getDataAssesmentText()
}
\value{
help text for data assesment
}
\description{
DataAssesment text
}
\examples{
    x<- getDataAssesmentText()

}
