% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{getDataMongo}
\alias{getDataMongo}
\title{getDataMongo}
\usage{
getDataMongo(
  dprofilerMongo = NULL,
  input = NULL,
  output = NULL,
  highlight = TRUE
)
}
\arguments{
\item{dprofilerMongo}{Dprofiler Mongo connection}

\item{input}{input}

\item{output}{output}

\item{highlight}{choose if you wanna highlight scores}
}
\description{
get Phenotypic data from the MongoDB server
}
\examples{
      x <- getDataMongo()
    
}
