% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getDeconvoluteTableDetails}
\alias{getDeconvoluteTableDetails}
\title{getDeconvoluteTableDetails}
\usage{
getDeconvoluteTableDetails(
  output = NULL,
  session = NULL,
  tablename = NULL,
  data = NULL,
  modal = NULL,
  highlight = FALSE
)
}
\arguments{
\item{output, }{output}

\item{session, }{session}

\item{tablename, }{table name}

\item{data, }{matrix data}

\item{modal, }{if it is true, the matrix is going to be in a modal}

\item{highlight}{if it is true, numerical columns are highlighted}
}
\value{
panel
}
\description{
get table details
To be able to put a table into two lines are necessary;
into the server part;
getDeconvoluteTableDetails(output, session, "dataname", data, modal=TRUE)
into the ui part;
uiOutput(ns("dataname"))
}
\examples{
    x <- getDeconvoluteTableDetails()

}
