% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{getDEgenesDownloadButtons}
\alias{getDEgenesDownloadButtons}
\title{getDEgenesDownloadButtons}
\usage{
getDEgenesDownloadButtons(
  output = NULL,
  session = NULL,
  DEgenes = NULL,
  IterDEgenes = NULL
)
}
\arguments{
\item{output}{output}

\item{session}{session}

\item{DEgenes}{Initial DE genes}

\item{IterDEgenes}{Final DE genes}
}
\description{
Buttons for downloading Initial, overlapping and DE genes
}
\examples{
    x <- getDEgenesDownloadButtons()
    
}
