% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getHeatmapMarkerGenes}
\alias{getHeatmapMarkerGenes}
\title{getHeatmapMarkerGenes}
\usage{
getHeatmapMarkerGenes(sc_marker_table = NULL, data = NULL, input = NULL)
}
\arguments{
\item{sc_marker_table}{single cell marker table}

\item{data}{reference bulk data}

\item{input}{input}
}
\description{
getHeatmapMarkerGenes
}
\examples{
    x <- getHeatmapMarkerGenes()
    
}
