% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getIconLabel}
\alias{getIconLabel}
\title{getIconLabel}
\usage{
getIconLabel(label = NULL, message = NULL)
}
\arguments{
\item{label}{label}

\item{message}{message}
}
\description{
creates a label with information icon
}
