% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{getIterDESummary}
\alias{getIterDESummary}
\title{getIterDESummary}
\usage{
getIterDESummary(
  output = NULL,
  session = NULL,
  vennname = NULL,
  summaryname = NULL,
  deres = NULL,
  params = NULL
)
}
\arguments{
\item{output}{output}

\item{session}{session}

\item{vennname}{venn diagram name}

\item{summaryname}{summary table name}

\item{deres}{DE results}

\item{params}{DE parameters}
}
\description{
get summary of Iterative DE Analysis results and venn diagram
}
\examples{
     x <- getIterDESummary()
    
}
