% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getIterMethodDetails}
\alias{getIterMethodDetails}
\title{getIterMethodDetails}
\usage{
getIterMethodDetails(num = 0, session = NULL, input = NULL)
}
\arguments{
\item{num, }{panel that is going to be shown}

\item{input, }{user input}
}
\description{
get the detail boxes after DE method selected
}
\examples{
    x <- getIterMethodDetails()

}
