% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getJSLine}
\alias{getJSLine}
\title{getJSLine}
\usage{
getJSLine()
}
\value{
JS Code
}
\description{
heatmap JS code for selection functionality. Adapted from debrowser::getJSLine()
}
\examples{
    x <- getJSLine()
    
}
