% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getLegendColors}
\alias{getLegendColors}
\title{getLegendColors}
\usage{
getLegendColors(Legend = c("Up", "Down", "NS"))
}
\arguments{
\item{Legend}{legend}
}
\description{
Generates colors according to the data. Adapted from debrowser::getLegendColors().
}
\examples{
      x <- getLegendColors()
    
}
