% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getMainPlot}
\alias{getMainPlot}
\title{getMainPlot}
\usage{
getMainPlot(
  input = NULL,
  output = NULL,
  session = NULL,
  mainname = NULL,
  width = NULL,
  plotdata = NULL,
  which_genes = NULL,
  DEgenes = NULL,
  IterDEgenes = NULL
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{mainname}{main plot name}

\item{width}{shiny box width}

\item{plotdata}{plot data}

\item{which_genes}{if true, DEgenes and IterDEgenes will be used for subseting}

\item{DEgenes}{list of initial DE genes}

\item{IterDEgenes}{list of Final DE genes}
}
\description{
a wrapper function for all main plots in Dprofiler
}
\examples{
     x <- getMainPlot()
    
}
