% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{getMarkerGenes}
\usage{
getMarkerGenes(scdata, IterDEgenes, input)
}
\arguments{
\item{scdata}{single cell data}

\item{IterDEgenes}{DE genes from bulk data}

\item{input}{input}
}
\description{
getMarkerGenes
}
