% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getMethodDetails_old}
\alias{getMethodDetails_old}
\title{get the detail boxes after DE method selected}
\usage{
getMethodDetails_old(num = 0, session = NULL, input = NULL)
}
\arguments{
\item{num}{panel that is going to be shown}

\item{session}{session}

\item{input}{user input}
}
\description{
get the detail boxes after DE method selected
}
\examples{
    x <- getMethodDetails()
    
}
