% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{getOutlierScores.old}
\alias{getOutlierScores.old}
\title{getOutlierScores.old}
\usage{
getOutlierScores.old(
  data = NULL,
  columns = NULL,
  conds = NULL,
  threshold = 0.5
)
}
\arguments{
\item{data}{data}

\item{columns}{samples}

\item{conds}{conditions}

\item{threshold}{threshold for score}

\item{deres}{DE results}
}
\description{
Calculate membership scores given the iterative DE results
}
\examples{
     x <- getOutlierScores()
    
}
