% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{getOutlierScores}
\alias{getOutlierScores}
\title{getOutlierScores}
\usage{
getOutlierScores(
  data = NULL,
  columns = NULL,
  conds = NULL,
  threshold = 0.5,
  pcs = 3,
  top = 2
)
}
\arguments{
\item{data}{data}

\item{columns}{samples}

\item{conds}{conditions}

\item{threshold}{threshold for score}

\item{pcs}{number of PCs}

\item{deres}{DE results}

\item{ntopgenes}{number of top genes}
}
\description{
Calculate membership scores given the iterative DE results
}
\examples{
     x <- getOutlierScores()
    
}
