% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getPlotArea}
\alias{getPlotArea}
\title{getPlotArea}
\usage{
getPlotArea(input = NULL, session = NULL)
}
\arguments{
\item{input}{}

\item{session}{}
}
\value{

}
\description{
a version of debrowser's getPlotArea function with automatic width and length
}
