% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{getProfileSampleDetails}
\alias{getProfileSampleDetails}
\title{getProfileSampleDetails}
\usage{
getProfileSampleDetails(
  output = NULL,
  summary = NULL,
  details = NULL,
  data = NULL
)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{details}{details output name}

\item{data}{data}
}
\value{

}
\description{
get single cell RNA samples details
}
\examples{

}
