% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfileScores}
\alias{getProfileScores}
\title{getProfileScores}
\usage{
getProfileScores(
  data = NULL,
  metadata = NULL,
  degenes = NULL,
  profiledata = NULL,
  profilemetadata = NULL,
  input = NULL
)
}
\arguments{
\item{data}{Bulk data}

\item{degenes}{DE genes}

\item{profiledata}{expression matrix of the profiling data}

\item{profilemetadata}{metadata of the profiling data}

\item{input}{input}
}
\description{
Calculate the profiling scores
}
\examples{
     x <- getProfileScores()

}
