% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getProfileSelector}
\alias{getProfileSelector}
\title{getProfileSelector}
\usage{
getProfileSelector(
  metadata = NULL,
  marker_table = NULL,
  session = NULL,
  input = NULL,
  n = 0
)
}
\arguments{
\item{metadata}{meta data table}

\item{marker_table}{marker_table}

\item{session}{session}

\item{input}{input params}

\item{n}{the box number}
}
\description{
Return the series selection box using profiling meta data table.
}
\examples{
     x <- getProfileSelector()
    
}
