% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfilingDEparameter}
\alias{getProfilingDEparameter}
\title{getProfilingDEparameter}
\usage{
getProfilingDEparameter(
  data = NULL,
  metadata = NULL,
  session = NULL,
  input = NULL
)
}
\arguments{
\item{data}{data}

\item{metadata}{metadata}

\item{session}{session}

\item{input}{input}
}
\description{
Prepare DE analysis parameters for the profiling data
}
\examples{
     x <- getProfilingDEparameter()
     
}
