% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getProfilingMethodDetails}
\alias{getProfilingMethodDetails}
\title{getProfilingMethodDetails}
\usage{
getProfilingMethodDetails(num = 0, session = NULL, input = NULL)
}
\arguments{
\item{num}{panel that is going to be shown}

\item{session}{session}

\item{input}{user input}
}
\description{
get the detail boxes after interative DE method selected
}
\examples{
    x <- getProfilingMethodDetails()

}
