% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfilingParameter}
\alias{getProfilingParameter}
\title{getProfilingParameter}
\usage{
getProfilingParameter(session = NULL, input = NULL)
}
\arguments{
\item{session}{session}

\item{input}{input}
}
\description{
Prepare DE analysis parameters for the profiling data
}
\examples{
     x <- getProfilingParameter()
     
}
