% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfilingUI}
\alias{getProfilingUI}
\title{getProfilingUI}
\usage{
getProfilingUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\description{
Creates a panel to visualize Profiling results
}
\examples{
    x <- getProfilingUI("profiling")

}
