% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputs.R
\name{getReferenceProfileRNA}
\alias{getReferenceProfileRNA}
\title{getReferenceProfileRNA}
\usage{
getReferenceProfileRNA(
  data,
  meta.data,
  name,
  sample_id = NULL,
  batch = NULL,
  secondary_batch = NULL,
  subset_profiledata = NULL,
  DEfield = NULL,
  DEparams = c("DESeq2", "parametric", FALSE, "Wald", "None", "Silhouette", "auto",
    "Log2FC+Padj", 1, 0.05, ""),
  max_count = 10
)
}
\arguments{
\item{data}{profiling data}

\item{meta.data}{profiling metadata}

\item{name}{name of the Bulk Reference}

\item{sample_id}{the field in the metadata that identifies a sample/biosample, if null, the first field will be the sample_id}

\item{batch}{the batch term from the meta.data}

\item{secondary_batch}{specify is there exists a secondary batch within the primary batch, e.g. Patient, or Donor}

\item{subset_profiledata}{a list of subsets based on meta.data fields, the names of each element should correspond to a field in the meta.data}

\item{DEfield}{the meta.data field to indicate a comparison, this should be a list with one element that this element should be a vector of two characters indicating the conditions}

\item{DEparams}{DE Analysis and Computational Profiling parameters}

\item{max_count}{remove genes whose maximum count is below a small count (e.g. 10)}
}
\description{
Prepares Seurat objects for Compositional Profiling
}
