% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputs.R
\name{getReferenceSingleCellRNA}
\alias{getReferenceSingleCellRNA}
\title{getReferenceSingleCellRNA}
\usage{
getReferenceSingleCellRNA(
  object,
  name,
  marker_ident = NULL,
  subsetProp = NULL,
  subset_seed = 1,
  subset_ident = NULL,
  subset_ident_min = 10,
  subset_assay = "RNA",
  subset_data = "counts",
  subset_embedding = "tsne"
)
}
\arguments{
\item{object}{Seurat Object}

\item{name}{name of the Single Cell Reference}

\item{marker_ident}{Ident of Seurat Object for marker analysis}

\item{subsetProp}{Proportion of single cell data to downsize, if NULL the entire single cell data is used}

\item{subset_seed}{The seed for random sampling}

\item{subset_ident}{Ident to subset the single cell data, if NULL cells are randomly selected}

\item{subset_ident_min}{Minimum number of cells in an ident, ignored if subset_ident is NULL}

\item{subset_assay}{Seurat assay used in the subset}

\item{subset_data}{Seurat data used in the subset}

\item{subset_embedding}{Seurat embedding used in the subset}
}
\description{
Prepares Seurat objects for Compositional Profiling
}
\examples{
# library
library(Seurat)

# prepare scRNA reference data
referencePBMC <- getReferenceSingleCellRNA(pbmc_small)
                                           
# prepare scRNA reference data, select idents for markers analysis and subsetting                                          
referencePBMC <- getReferenceSingleCellRNA(pbmc_small, marker_ident = c("RNA_snn_res.0.8","RNA_snn_res.1"), 
                                           subsetProp = 0.8, subset_ident = "RNA_snn_res.0.8", subset_ident_min = 10)

}
