% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{getScoreDetails}
\alias{getScoreDetails}
\title{getScoreDetails}
\usage{
getScoreDetails(
  output = NULL,
  session = NULL,
  plotname = NULL,
  DEscores = NULL,
  OutlierDEscores = NULL
)
}
\arguments{
\item{output}{output}

\item{session}{session}

\item{plotname}{the name of the plot}

\item{DEscores}{Cross Condition Scores of impure conditions}

\item{OutlierDEscores}{Intra Condition Scores of impure conditions}

\item{IterDEscores}{Cross Condition Scores of pure conditions}

\item{OutlierIterDEscores}{Intra Condition Scores of pure conditions}
}
\description{
get score details of Iterative DE analysis
}
\examples{
     x <- getScoreDetails()
    
}
