% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getSCRNAEmbedding}
\alias{getSCRNAEmbedding}
\title{getSCRNAEmbedding}
\usage{
getSCRNAEmbedding(output = NULL, summary = NULL, data = NULL, ident = NULL)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{data}{single cell ExpressionSet Object}

\item{ident}{column in scRNA metadata to visualize}
}
\description{
get single cell RNA Embedding
}
\examples{
      x <- getSCRNAEmbedding()
    
}
