% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getScRNAMetaSelector}
\alias{getScRNAMetaSelector}
\title{getScRNAMetaSelector}
\usage{
getScRNAMetaSelector(metadata = NULL, session = NULL, input = NULL, n = 0)
}
\arguments{
\item{metadata}{meta data table}

\item{session}{session}

\item{input}{input params}

\item{n}{the box number}
}
\description{
Return the sample selection box using meta data table of scRNA data.
}
\examples{
     x <- getScRNAMetaSelector()
    
}
