% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{getSCRNASampleDetails}
\alias{getSCRNASampleDetails}
\title{getSCRNASampleDetails}
\usage{
getSCRNASampleDetails(
  output = NULL,
  summary = NULL,
  details = NULL,
  data = NULL,
  ident = NULL,
  UMI_column = NULL
)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{details}{details output name}

\item{data}{single cell ExpressionSet Object}
}
\value{

}
\description{
get single cell RNA samples details
}
