% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getSeriesSelector}
\alias{getSeriesSelector}
\title{getSeriesSelector}
\usage{
getSeriesSelector(metadata = NULL, session = NULL, input = NULL, n = 0)
}
\arguments{
\item{metadata}{meta data table}

\item{session}{session}

\item{input}{input params}

\item{n}{the box number}
}
\description{
Return the series selection box using profiling meta data table.
}
\examples{
     x <- getSeriesSelector()
    
}
