% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getTabUpdateJS}
\alias{getTabUpdateJS}
\title{getTabUpdateJS}
\usage{
getTabUpdateJS()
}
\value{
the JS for tab updates
}
\description{
prepmenu tab and discovery menu tab updates, adapted from debrowser::getTabUpdateJS()
}
\examples{
    x<- getTabUpdateJS()

}
