% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapfunctions.R
\name{heatmapServer}
\alias{heatmapServer}
\title{heatmapServer}
\usage{
heatmapServer(input = NULL, output = NULL, session = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Sets up shinyServer to be able to run heatmapServer interactively.
Adapted from debrowser::heatmapServer()
}
\examples{
    x <- heatmapServer()

}
