% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapfunctions.R
\name{heatmapUI}
\alias{heatmapUI}
\title{heatmapUI}
\usage{
heatmapUI(input = NULL, output = NULL, session = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
the panel for heatmapUI;
}
\description{
Creates a shinyUI to be able to run DEBrowser interactively.
}
\examples{
    x<-heatmapUI()

}
