% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{insertDataMongo}
\alias{insertDataMongo}
\title{insertDataMongo}
\usage{
insertDataMongo(
  dprofilerMongo = NULL,
  output = NULL,
  table = NULL,
  dataname = NULL
)
}
\arguments{
\item{dprofilerMongo}{Dprofiler Mongo connection}

\item{output}{output}

\item{table}{MongoDB database sample table}

\item{dataname}{Optional, ask for the name (Series) of the newly inserted dataset}

\item{tablename}{Table name in shiny}
}
\description{
insert Phenotypic data to the MongoDB server
}
\examples{
      x <- insertDataMongo()
    
}
