% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{MainLoadPage}
\alias{MainLoadPage}
\title{MainLoadPage}
\usage{
MainLoadPage()
}
\description{
Create tab items for the Main Page which includes the data loading modules (bulk, single cell, bulk reference), 
and mongo server panes
}
