% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{MergeScoreTables}
\alias{MergeScoreTables}
\title{MergeScoreTables}
\usage{
MergeScoreTables(CrossScore = NULL, IntraScore = NULL)
}
\arguments{
\item{CrossScore}{Cross condition score tables}

\item{IntraScore}{Intra condition score tables}
}
\description{
MergeScoreTables
}
\examples{
    x <- MergeScoreTables()
    
}
