% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{mongoServer}
\alias{mongoServer}
\title{mongoServer}
\usage{
mongoServer(
  input = NULL,
  output = NULL,
  session = NULL,
  parent_session = NULL,
  uploadeddata = NULL,
  summarisedData = NULL
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{uploadeddata}{uploaded data from initial dprofiler module}

\item{comppheno}{results from computational phenotypic profiling}

\item{composprof}{results from cmopositional profiling}

\item{comparprof}{results from comparative profiling}
}
\description{
a module server to manage mongo server
}
\examples{
    x <- mongoServer()
    
}
