% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{mongoServerUI}
\alias{mongoServerUI}
\title{mongoServerUI}
\usage{
mongoServerUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Creates a side panel for server set up!
}
\examples{
    x <- mongoServerUI()
    
}
