% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{outlier_silhouette}
\alias{outlier_silhouette}
\title{outlier_silhouette}
\usage{
outlier_silhouette(cluster = NULL, dist2 = NULL, dist = NULL)
}
\arguments{
\item{cluster}{cluster labels of profile data}

\item{dist2}{non-symmetric similarity matrix}

\item{dist}{symmmetric similarity matrix}
}
\description{
Calculate outlier scores for non-symmetric distance matrices. 
Suitable for comparison of expression profiles across different datasets and experiments 
detecing outlying samples compared to reference bulk data sets
}
\examples{
     x <- outlier_silhouette()
    
}
