% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{prepDataContainer}
\alias{prepDataContainer}
\title{prepDataContainer}
\usage{
prepDataContainer(data = NULL, input = NULL, session = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{input, }{input parameters}

\item{session}{session}
}
\description{
Prepares the data container that stores values used within DE analysis. Adapted from debrowser::prepDataContainer.
}
\examples{
    x <- prepDataContainer()

}
