% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{prepDataContainer_old}
\alias{prepDataContainer_old}
\title{prepDataContainer}
\usage{
prepDataContainer_old(input = NULL, session = NULL, data = NULL, meta = NULL)
}
\arguments{
\item{input, }{input parameters}

\item{session}{session}

\item{data, }{loaded dataset}

\item{meta, }{loaded metadata}
}
\description{
Prepares the data container that stores values used within DE analysis. Adapted from debrowser::prepDataContainer.
OLD FUNCTION DELETE LATER
}
\examples{
    x <- prepDataContainer()
    
}
