% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{prepDeconvolute}
\alias{prepDeconvolute}
\title{prepDeconvolute}
\usage{
prepDeconvolute(dc = NULL, scdata = NULL, parent_session = NULL)
}
\arguments{
\item{dc}{reactive object of DE analysis}

\item{scdata}{single cell ExpressionSet Object}

\item{parent_session}{main session}
}
\value{

}
\description{
Prepares the container for deconvolution methods
}
