% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{prepDeconvolute_old}
\alias{prepDeconvolute_old}
\title{prepDeconvolute}
\usage{
prepDeconvolute_old(
  dc = NULL,
  scdata = NULL,
  sc_marker_table = NULL,
  parent_session = NULL
)
}
\arguments{
\item{dc}{reactive object of DE analysis}

\item{scdata}{single cell ExpressionSet Object}

\item{sc_marker_table}{single cell markers table}

\item{parent_session}{parent session}
}
\description{
Prepares the container for deconvolution methods
}
\examples{
       x <- prepDeconvolute()
    
}
