% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{prepDEparameters}
\alias{prepDEparameters}
\title{prepDEparameters}
\usage{
prepDEparameters(input = NULL, session = NULL)
}
\arguments{
\item{input, }{input parameters}

\item{session}{session}
}
\description{
Prepares the parameters for DE analysis. Adapted from debrowser::prepDataContainer.
}
\examples{
    x <- prepDEparameters()
    
}
