% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{profilefileUploadBox}
\alias{profilefileUploadBox}
\title{profilefileUploadBox}
\usage{
profilefileUploadBox(
  id = NULL,
  inputId_select = NULL,
  inputId_external = NULL,
  inputId_button = NULL,
  inputId_dirbutton = NULL,
  label = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{inputId_select}{input ID for selecting references}

\item{inputId_external}{input ID for custom external reference}

\item{inputId_button}{input ID for import button}

\item{inputId_dirbutton}{input ID for directory select button}

\item{label}{label}
}
\description{
file upload box for profile bulk data and metadata. Adapted from debrowser::fileUploadBox()
}
\examples{
      x <- profilefileUploadBox("meta", "count", "metadata", "Metadata")
    
}
