% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{runMultipleDE}
\alias{runMultipleDE}
\title{runMultipleDE}
\usage{
runMultipleDE(data = NULL, columns = NULL, conds = NULL, params = NULL)
}
\arguments{
\item{data}{A matrix that includes all the expression raw counts, rownames has to be the gene, isoform or region names/IDs.}

\item{columns}{a vector that includes the columns that are going to be analyzed. These columns has to match with the given data.}

\item{conds}{Multiple experimental conditions.}

\item{params}{all params for the DE methods}
}
\description{
Run DE algorithms on the selected parameters. Output is to be used for the interactive display.
Multiple comparisons version of runDE.
}
\examples{
     x <- runMultipleDE()
     
}
