% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{scfileUploadBox}
\alias{scfileUploadBox}
\title{scfileUploadBox}
\usage{
scfileUploadBox(
  id = NULL,
  inputId_select = NULL,
  inputId_external = NULL,
  inputId_button = NULL,
  inputId_dirbutton = NULL,
  label = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{inputId_select}{input ID for selecting references}

\item{inputId_external}{input ID for custom external reference}

\item{inputId_button}{input ID for import button}

\item{inputId_dirbutton}{input ID for directory select button}

\item{label}{label}
}
\description{
file upload box for single cell ExpressionSet object.
}
\examples{
      x <- scfileUploadBox("meta", "metadata", "Metadata")
    
}
