% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{ScoreCutOffSelectionUI}
\alias{ScoreCutOffSelectionUI}
\title{ScoreCutOffSelectionUI}
\usage{
ScoreCutOffSelectionUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the left menu according to the selected tab;
}
\description{
Gathers the cut off selection for Scoring
}
\note{
\code{ScoreCutOffSelectionUI}
}
\examples{
    x <- ScoreCutOffSelectionUI("cutoff")
}
