% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{selectDataMongo}
\alias{selectDataMongo}
\title{selectDataMongo}
\usage{
selectDataMongo(
  dprofilerMongo = NULL,
  input = NULL,
  session = NULL,
  newSeries = NULL
)
}
\arguments{
\item{dprofilerMongo}{Dprofiler Mongo connection}

\item{output}{output}

\item{table}{MongoDB database sample table}
}
\description{
select a series from the search box of Series in the Mongo Server
}
\examples{
      x <- selectDataMongo()
    
}
