% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectScRNAConditions}
\alias{selectScRNAConditions}
\title{selectScRNAConditions}
\usage{
selectScRNAConditions(scdata = NULL, session = NULL, input = NULL)
}
\arguments{
\item{scdata, }{used single cell dataset}

\item{session, }{session}

\item{input, }{input params}

\item{metadata, }{metadatatable to select from metadata}
}
\description{
Selects user input conditions, multiple if present, to be
used in DESeq.
}
\examples{
    x<- selectScRNAConditions()

}
