% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{sepRadio}
\alias{sepRadio}
\title{sepRadio}
\usage{
sepRadio(id, name)
}
\arguments{
\item{id}{module id}

\item{name}{}
}
\value{

}
\description{
Radio button for seperators: an inline version of debrowser's sepRadio function
}
