% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{simulateProfilingData}
\alias{simulateProfilingData}
\title{simulateProfilingData}
\usage{
simulateProfilingData(compcodeR.param = NULL, overlappingDE = NULL, seed = 1)
}
\arguments{
\item{compcodeR.param}{parameters for the compcodeR package, simulating reference and target data}

\item{overlappingDE}{the percentage of the number of overlapping DE genes for reference and target data}

\item{seed}{seed value for random number generation}
}
\description{
simulates a reference and target data for testing Comparative Profiling
}
\examples{
      x <- simulateProfilingData()

}
