% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startDprofiler}
\alias{startDprofiler}
\title{startDprofiler}
\usage{
startDprofiler()
}
\value{
the app
}
\description{
Starts the Dprofiler to be able to run interactively.
}
\note{
\code{startDprofiler}
}
\examples{
    startDprofiler()

}
