% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{SummaryUI}
\alias{SummaryUI}
\title{SummaryUI}
\usage{
SummaryUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Creates the Mongo Database Pane
}
\examples{
    x <- SummaryUI("summariser")
    
}
