% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoserver.R
\name{UpdateMongoSearchBox}
\alias{UpdateMongoSearchBox}
\title{UpdateMongoSearchBox}
\usage{
UpdateMongoSearchBox(dprofilerMongo = NULL, input = NULL, session = NULL)
}
\arguments{
\item{dprofilerMongo}{Dprofiler Mongo connection}

\item{output}{output}

\item{table}{MongoDB database sample table}
}
\description{
updates the search box of Series in the Mongo Server
}
\examples{
      x <- UpdateMongoSearchBox()
    
}
