% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{validatebulkReference}
\alias{validatebulkReference}
\title{validatebulkReference}
\usage{
validatebulkReference(import_path, session)
}
\arguments{
\item{import_path}{path to the bulk reference directory}

\item{session}{session}
}
\description{
Validates if the bulk reference folder has all necessary files
}
