% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{validatescReference}
\alias{validatescReference}
\title{validatescReference}
\usage{
validatescReference(import_path, session)
}
\arguments{
\item{import_path}{path to the single cell reference directory}

\item{session}{session}
}
\description{
Validates if the single cell refernece folder has all necessary files
}
