% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{custom_silhouette}
\alias{custom_silhouette}
\title{custom_silhouette}
\usage{
custom_silhouette(x = NULL, dist = NULL)
}
\arguments{
\item{x}{conditions}

\item{dist}{distance matrix}
}
\description{
A custom silhouette function for calculating the silhouette measure of each point with respect
to each condition
}
\examples{
     x <- custom_silhouette()
     
}
