% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{deconvolute}
\usage{
deconvolute(data = NULL, DEgenes = NULLL, scdata = NULL, input = NULL)
}
\arguments{
\item{data}{Bulk expression data set}

\item{DEgenes}{DE genes for limiting the genes of scRNA and Bulk RNA data sets}

\item{scdata}{single cell ExpressionSet Object}

\item{input}{input}
}
\description{
the deconvolution function based on MuSiC algorithm
}
\examples{
    x <- deconvolute()
    
}
