% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{display_venn}
\alias{display_venn}
\title{display_venn}
\usage{
display_venn(x = NULL, category.names = NULL, ...)
}
\arguments{
\item{x}{a list of elements for each groups}

\item{category.names}{names of venn diagram categories}

\item{...}{addtional parameters passed to draw.pairwise.venn function.}
}
\description{
display venn diagram
}
\examples{
     x <- display_venn()
     
}
