% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{dprofilerdeanalysis}
\alias{dprofilerdeanalysis}
\title{dprofilerdeanalysis}
\usage{
dprofilerdeanalysis(
  input = NULL,
  output = NULL,
  session = NULL,
  data = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL,
  parent_session = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}

\item{columns, }{columns}

\item{conds, }{conditions}

\item{params, }{de parameters}

\item{parent_session}{parent session}
}
\description{
Module to perform and visualize Iterative DE results.
}
\examples{
    x <- dprofilerdeanalysis()
    
}
