% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{dprofilerdeconvolute}
\alias{dprofilerdeconvolute}
\title{dprofilerdeconvolute}
\usage{
dprofilerdeconvolute(
  input = NULL,
  output = NULL,
  session = NULL,
  parent_session = NULL,
  data = NULL
)
}
\arguments{
\item{input}{input variables}

\item{output}{output objects}

\item{session}{session}

\item{parent_session}{parent session}

\item{dc}{de results}

\item{scdata}{single cell data}

\item{sc_marker_table}{single cell marker table}
}
\value{
DE panel
}
\description{
Module to perform and visualize deconvolution results.
}
\examples{
    x <- dprofilerdeconvolute() 
    
}
