% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{dprofilerprofiling}
\alias{dprofilerprofiling}
\title{dprofilerprofiling}
\usage{
dprofilerprofiling(
  input = NULL,
  output = NULL,
  session = NULL,
  dc = NULL,
  profiledata = NULL,
  profilemetadata = NULL,
  parent_session = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{dc, }{results of iterative DE Analysis}

\item{profiledata}{expression matrix of profile data}

\item{profilemetadata}{metadata of profile data}

\item{parent_session}{parent session}
}
\description{
Module to perform and visualize deconvolution results.
}
\examples{
    x <- dprofilerprofiling()
    
}
