% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getAllMarkerGenes}
\alias{getAllMarkerGenes}
\title{getAllMarkerGenes}
\usage{
getAllMarkerGenes(
  sc_marker_table = NULL,
  scdata = NULL,
  data = NULL,
  input = NULL
)
}
\arguments{
\item{sc_marker_table}{single cell marker table}

\item{scdata}{single cell data object}

\item{data}{bulk data}

\item{input}{input}

\item{columns}{columns of the bulk data}
}
\description{
getAllMarkerGenes
}
\examples{
    x <- getAllMarkerGenes()
    
}
