% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{getDeconvoluteUI}
\alias{getDeconvoluteUI}
\title{getDeconvoluteUI}
\usage{
getDeconvoluteUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
Creates a panel to visualize deconvolution results
}
\examples{
    x <- getDeconvoluteUI("batcheffect")

}
