% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getExpressionProfiles}
\alias{getExpressionProfiles}
\title{Title}
\usage{
getExpressionProfiles(deres = NULL, data = NULL, columns = NULL, conds = NULL)
}
\arguments{
\item{deres}{}

\item{data}{}

\item{columns}{}

\item{conds}{}
}
\value{

}
\description{
Given the gene expression matrix with multiple conditions, generate expression profiles.
}
