% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{getFinalScores}
\alias{getFinalScores}
\title{getFinalScores}
\usage{
getFinalScores(
  deres = NULL,
  data = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL,
  ManualDEgenes = NULL,
  TopStat = NULL
)
}
\arguments{
\item{deres}{DE results}

\item{data}{data}

\item{columns}{samples}

\item{conds}{conditions}

\item{params}{DE params}

\item{ManualDEgenes}{manually inserted list of DE genes}

\item{TopStat}{the number of top statistics}
}
\description{
Calculate membership scores given the iterative DE results
}
\examples{
     x <- getFinalScores()
    
}
