% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getIdentSelectorFromMeta}
\alias{getIdentSelectorFromMeta}
\title{getIdentSelectorFromMeta}
\usage{
getIdentSelectorFromMeta(
  metadata = NULL,
  session = NULL,
  input = NULL,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{metadata}{meta data table}

\item{session}{session}

\item{input}{input}

\item{choices}{choices}

\item{selected}{selected}
}
\description{
Select identification from single cell metadata.
}
\examples{
     x <- getIdentSelectorFromMeta()
     
}
