% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getIterMethodDetails}
\alias{getIterMethodDetails}
\title{getIterMethodDetails}
\usage{
getIterMethodDetails(num = 0, input = NULL)
}
\arguments{
\item{num}{panel that is going to be shown}

\item{input}{user input}

\item{session}{session}
}
\description{
get the detail boxes after interative DE method selected
}
\examples{
    x <- getIterMethodDetails()
    
}
