% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{getProfileSampleDetails}
\alias{getProfileSampleDetails}
\title{getProfileSampleDetails}
\usage{
getProfileSampleDetails(
  output = NULL,
  summary = NULL,
  details = NULL,
  data = NULL,
  ident1 = NULL,
  ident2 = NULL
)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{details}{details output name}

\item{data}{data}

\item{ident1}{column in profile metadata for PCA plot 1}

\item{ident2}{column in profile metadata for PCA plot 2}
}
\description{
get Profile Data Sample Details
}
\examples{
     x <- getProfileSampleDetails()
    
}
