% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfileScoreDetails}
\alias{getProfileScoreDetails}
\title{getProfileScoreDetails}
\usage{
getProfileScoreDetails(
  output = NULL,
  session = NULL,
  plotname = NULL,
  scores = NULL
)
}
\arguments{
\item{output}{output}

\item{session}{session}

\item{plotname}{name of the plot}

\item{scores}{profiling scores}
}
\description{
Details and scores of the profiling analysis
}
\examples{
    x <- getProfileScoreDetails()
    
}
