% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfileScores}
\alias{getProfileScores}
\title{getProfileScores}
\usage{
getProfileScores(
  dc = NULL,
  profiledata = NULL,
  profilemetadata = NULL,
  params = NULL
)
}
\arguments{
\item{dc}{DE results}

\item{profiledata}{expression matrix of the profiling data}

\item{profilemetadata}{metadata of the profiling data}

\item{params}{profiling parameters}
}
\description{
Calculate the profiling scores
}
\examples{
     x <- getProfileScores()
    
}
