% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{getProfileScores}
\alias{getProfileScores}
\title{getProfileScores}
\usage{
getProfileScores(
  dc = NULL,
  dat = NULL,
  profiledata = NULL,
  profilemetadata = NULL
)
}
\arguments{
\item{dc}{DE results}

\item{dat}{expression matrix of the main data}

\item{profiledata}{expression matrix of the profiling data}

\item{profilemetadata}{metadata of the profiling data}
}
\description{
Calculate the profiling scores
}
\examples{
     x <- getProfileScores()

}
