% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{getScoreDetails}
\alias{getScoreDetails}
\title{getScoreDetails}
\usage{
getScoreDetails(
  output = NULL,
  session = NULL,
  plotname = NULL,
  DEscores = NULL,
  IterDEscores = NULL
)
}
\arguments{
\item{output}{output}

\item{session}{session}

\item{plotname}{the name of the plot}

\item{DEscores}{Scores of impure conditions}

\item{IterDEscores}{Scores of pure conditions}
}
\value{

}
\description{
get score details of columns
}
