% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{getSCRNASampleDetails}
\alias{getSCRNASampleDetails}
\title{getSCRNASampleDetails}
\usage{
getSCRNASampleDetails(
  output = NULL,
  summary = NULL,
  details = NULL,
  data = NULL,
  ident = NULL,
  UMI_column = NULL
)
}
\arguments{
\item{output}{output}

\item{summary}{summary output name}

\item{details}{details output name}

\item{data}{single cell ExpressionSet Object}

\item{ident}{column in scRNA metadata to visualize}

\item{UMI_column}{column in scRNA metadata with total UMI counts}
}
\description{
get single cell RNA samples details
}
\examples{
      x <- getSCRNASampleDetails()
      
}
