% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{profilefileUploadBox}
\alias{profilefileUploadBox}
\title{profilefileUploadBox}
\usage{
profilefileUploadBox(
  id = NULL,
  inputId_count = NULL,
  inputId_meta = NULL,
  inputId_dmeta = NULL,
  inputId_dmeta_key = NULL,
  label = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{inputId_count}{input data file ID}

\item{inputId_meta}{input metadata file ID}

\item{inputId_dmeta}{dmeta API URL}

\item{inputId_dmeta_key}{dmeta token}

\item{label}{label}
}
\description{
file upload box for profile bulk data and metadata. Adapted from debrowser::fileUploadBox()
}
\examples{
      x <- profilefileUploadBox("meta", "count", "metadata", "Metadata")

}
