% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterdeprogs.R
\name{runIterDE}
\alias{runIterDE}
\title{runIterDE}
\usage{
runIterDE(
  data = NULL,
  metadata = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL,
  session = NULL,
  verbose = FALSE,
  visualize_prefix = "plot"
)
}
\arguments{
\item{data, }{A matrix that includes all the expression raw counts,
rownames has to be the gene, isoform or region names/IDs}

\item{metadata, }{metadata}

\item{columns, }{is a vector that includes the columns that are going
to be analyzed. These columns has to match with the given data.}

\item{conds, }{experimental conditions. The order has to match
with the column order}

\item{params, }{all params for the DE methods}

\item{session, }{the session}

\item{verbose, }{TRUE if you want results to be printed}

\item{visualize_prefix, }{the prefix of the PCA plot}
}
\description{
Run Iterative DE algorithms on the selected parameters.  Output is
to be used for the interactive display.
}
\examples{
    x <- runIterDE()
    
}
