% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{scfileUploadBox}
\alias{scfileUploadBox}
\title{scfileUploadBox}
\usage{
scfileUploadBox(id = NULL, inputId = NULL, label = NULL)
}
\arguments{
\item{id}{namespace id}

\item{label}{label}

\item{inputId_count}{input data file ID}

\item{inputId_meta}{input metadata file ID}
}
\value{

}
\description{
file upload module for single cell ExpressionSet object
}
