% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{scfileUploadBox}
\alias{scfileUploadBox}
\title{scfileUploadBox}
\usage{
scfileUploadBox(id = NULL, inputId = NULL, label = NULL)
}
\arguments{
\item{id}{namespace id}

\item{inputId}{input data file ID}

\item{label}{label}
}
\description{
file upload box for single cell ExpressionSet object.
}
\examples{
      x <- fileUploadBox("meta", "metadata", "Metadata")
    
}
