% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{ScoreCutOffSelectionUI}
\alias{ScoreCutOffSelectionUI}
\title{ScoreCutOffSelectionUI}
\usage{
ScoreCutOffSelectionUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Gathers the cut off selection for scoring
}
\examples{
    x <- ScoreCutOffSelectionUI("cutoff")
    
}
