% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectProfilingConditions}
\alias{selectProfilingConditions}
\title{selectProfilingConditions}
\usage{
selectProfilingConditions(
  Dataset = NULL,
  metadata = NULL,
  session = NULL,
  input = NULL
)
}
\arguments{
\item{Dataset, }{used dataset}

\item{metadata, }{metadatatable to select from metadata}

\item{session, }{session}

\item{input, }{input params}
}
\description{
Select metadata field to use DE results from reference data
}
\examples{
    x<- selectProfilingConditions()
    
}
