% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectScRNAConditions}
\alias{selectScRNAConditions}
\title{selectScRNAConditions}
\usage{
selectScRNAConditions(scdata = NULL, session = NULL, input = NULL)
}
\arguments{
\item{scdata, }{used single cell dataset}

\item{session, }{session}

\item{input, }{input params}

\item{choicecounter, }{choicecounter to add multiple comparisons}
}
\description{
Selects user input conditions used in bulk RNA deconvolution.
}
\examples{
    x<- selectScRNAConditions()
    
}
